refactor(models): move trade insights contracts
