+++ src/uw_scan/models/__init__.py
@@ -100,6 +100,43 @@ from .volatility import (
     VolStatsRow,
     VrpDailyPoint,
 )
+from .trade_insights import (
+    CandidateStructure,
+    ChainFlowReadRow,
+    InsightBadge,
+    InsightLeg,
+    InsightsSynthesis,
+    InsightSignalRow,
+    SourceReconciliation,
+    SourceReconciliationRow,
+    TermMoveRow,
+    TradeInsightsHeader,
+    TradeInsightsResponse,
+)
+from .trade_insights_ai import (
+    TradeInsightAiAnalysisRequest,
+    TradeInsightAiAnalysisResponse,
+    TradeInsightAiBase,
+    TradeInsightAiBestExpression,
+    TradeInsightAiConflict,
+    TradeInsightAiDominantRead,
+    TradeInsightAiGuardrails,
+    TradeInsightAiHeadline,
+    TradeInsightAiHighlight,
+    TradeInsightAiLevel,
+    TradeInsightAiMetricCard,
+    TradeInsightAiOutcome,
+    TradeInsightAiPreferredExpression,
+    TradeInsightAiRejectedIdea,
+    TradeInsightAiRendering,
+    TradeInsightAiRequiredCheck,
+    TradeInsightAiScenarioCard,
+    TradeInsightAiScoreBreakdown,
+    TradeInsightAiSectionCard,
+    TradeInsightAiSectionCards,
+    TradeInsightAiSnapshotMeta,
+    TradeInsightAiVrpAssessment,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -147,315 +184,9 @@ from .volatility import (
 # ---------------------------------------------------------------------------
 
 
-class InsightBadge(_UwBase):
-    code: str
-    label: str
-    severity: str = "info"
-
-
-class TradeInsightsHeader(_UwBase):
-    dominant_bias: str = "NEUTRAL"
-    primary_setup: str = "NO_CLEAR_SETUP"
-    confidence_label: str = "LOW"
-    data_quality_label: str = "INSUFFICIENT"
-    idea_count: int = 0
-    preferred_idea_id: str | None = None
-    badges: list[InsightBadge] = []
-
-
-class SourceReconciliationRow(_UwBase):
-    source_pair: str
-    price_agreement: str = ""
-    iv_agreement: str = ""
-    decision: str = ""
-    strike: Decimal | None = None
-    source_a_call_iv: Decimal | None = None
-    source_b_call_iv: Decimal | None = None
-    iv_diff: Decimal | None = None
-
-
-class SourceReconciliation(_UwBase):
-    status: str = "UNKNOWN"
-    headline: str = "Source reconciliation unavailable"
-    primary_iv_source: str | None = None
-    relative_shape_source: str | None = None
-    rows: list[SourceReconciliationRow] = []
-    decision: str = "Use deterministic data only where source agreement is understood."
-
-
-class InsightSignalRow(_UwBase):
-    lens: str
-    read: str
-    evidence: list[str] = []
-    conflicts: list[str] = []
-
-
-class ChainFlowReadRow(_UwBase):
-    strike: Decimal
-    call_volume: int | None = None
-    call_open_interest: int | None = None
-    put_volume: int | None = None
-    put_open_interest: int | None = None
-    call_put_volume_ratio: Decimal | None = None
-    volume_oi_note: str = ""
-    read: str = ""
-    requires_t1_oi_confirmation: bool = False
-
-
-class TermMoveRow(_UwBase):
-    expiry: _date
-    dte: int | None = None
-    atm_straddle: Decimal | None = None
-    implied_move_perc: Decimal | None = None
-    daily_implied_move_perc: Decimal | None = None
-    read: str = ""
-
-
-class InsightLeg(_UwBase):
-    side: str
-    option_symbol: str
-    option_right: str
-    expiry: _date
-    strike: Decimal
-    mid: Decimal | None = None
-
-
-class CandidateStructure(_UwBase):
-    idea_id: str
-    structure: str
-    thesis: str
-    expression_type: str
-    legs: list[InsightLeg] = []
-    net_credit_debit: Decimal | None = None
-    max_profit: Decimal | None = None
-    max_loss: Decimal | None = None
-    breakevens: list[Decimal] = []
-    profit_zone: str = ""
-    edge_source: str = ""
-    risk_flags: list[str] = []
-    rank: int
-    status: str = "candidate"
-
-
-class InsightsSynthesis(_UwBase):
-    dominant_story: str = ""
-    preferred_idea_id: str | None = None
-    best_risk_reward_idea_id: str | None = None
-    avoid: list[str] = []
-    required_before_sizing: list[str] = []
-
-
-class TradeInsightsResponse(_UwBase):
-    ticker: str
-    as_of: datetime | None = None
-    mode: str = "research"
-    header: TradeInsightsHeader
-    source_reconciliation: SourceReconciliation = SourceReconciliation()
-    signal_stack: list[InsightSignalRow] = []
-    flow_table: list[ChainFlowReadRow] = []
-    term_structure_table: list[TermMoveRow] = []
-    candidate_structures: list[CandidateStructure] = []
-    synthesis: InsightsSynthesis = InsightsSynthesis()
-
-
 # ---------------------------------------------------------------------------
 # Trade Insights AI analysis (V1.5)
 # ---------------------------------------------------------------------------
-class TradeInsightAiBase(BaseModel):
-    model_config = ConfigDict(extra="forbid", frozen=False)
-
-
-class TradeInsightAiDominantRead(TradeInsightAiBase):
-    headline: str
-    summary: str
-    confidence_commentary: str
-    data_quality_commentary: str
-
-
-class TradeInsightAiSnapshotMeta(TradeInsightAiBase):
-    run_id: int
-    trade_insights_input_hash: str
-    analysis_input_hash: str
-    data_as_of: str | None = None
-    freshness_label: str = "unknown"
-    source_notes: list[str] = Field(default_factory=list)
-
-
-class TradeInsightAiHeadline(TradeInsightAiBase):
-    title: str
-    stance: Literal["bullish", "bearish", "neutral", "mixed", "wait"]
-    stance_label: str
-    score: int
-    score_scale: int = 100
-    conviction: str
-    conviction_label: str
-    top_reason: str
-    primary_risk: str
-    watch_trigger: str
-
-
-class TradeInsightAiMetricCard(TradeInsightAiBase):
-    label: str
-    value: str
-    tone: str = "neutral"
-    source_path: str | None = None
-    note: str = ""
-
-
-class TradeInsightAiScenarioCard(TradeInsightAiBase):
-    case: Literal["upside", "base", "downside"] | str
-    tone: str = "neutral"
-    title: str
-    description: str
-
-
-class TradeInsightAiScoreBreakdown(TradeInsightAiBase):
-    section: str
-    score: int
-    max_score: int
-    summary: str
-
-
-class TradeInsightAiHighlight(TradeInsightAiBase):
-    label: str
-    value: str
-    source_path: str | None = None
-    note: str = ""
-
-
-class TradeInsightAiLevel(TradeInsightAiBase):
-    price: str
-    kind: str
-    value: str
-    importance: str = "normal"
-    source_path: str | None = None
-    note: str = ""
-
-
-class TradeInsightAiSectionCard(TradeInsightAiBase):
-    title: str
-    score: int | None = None
-    max_score: int | None = None
-    summary: str
-    highlights: list[TradeInsightAiHighlight] = Field(default_factory=list)
-    levels: list[TradeInsightAiLevel] = Field(default_factory=list)
-    data_quality: str = "unknown"
-
-
-class TradeInsightAiSectionCards(TradeInsightAiBase):
-    market_structure: TradeInsightAiSectionCard
-    volatility: TradeInsightAiSectionCard
-    flow_positioning: TradeInsightAiSectionCard
-
-
-class TradeInsightAiVrpAssessment(TradeInsightAiBase):
-    signal: str
-    title: str
-    summary: str
-    metrics: list[TradeInsightAiMetricCard] = Field(default_factory=list)
-    reason: str
-
-
-class TradeInsightAiPreferredExpression(TradeInsightAiBase):
-    idea_id: str
-    structure: str
-    title: str
-    subtitle: str = ""
-    estimated_entry: str = ""
-    max_profit_observed: str = ""
-    max_loss_observed: str = ""
-    reward_risk: str = ""
-    why: str
-    management_notes: list[str] = Field(default_factory=list)
-    status_observed: str
-    risk_flags_observed: list[str] = Field(default_factory=list)
-
-
-class TradeInsightAiBestExpression(TradeInsightAiBase):
-    idea_id: str
-    structure: str
-    role: str
-    why: str
-    caveats: list[str] = Field(default_factory=list)
-    status_observed: str
-    risk_flags_observed: list[str] = Field(default_factory=list)
-
-
-class TradeInsightAiConflict(TradeInsightAiBase):
-    lens: str
-    severity: str
-    description: str
-    affected_idea_ids: list[str] = Field(default_factory=list)
-
-
-class TradeInsightAiRequiredCheck(TradeInsightAiBase):
-    check: str
-    reason: str
-    blocks_sizing: bool = True
-    source: str = ""
-
-
-class TradeInsightAiRejectedIdea(TradeInsightAiBase):
-    idea_id: str
-    structure: str
-    reason: str
-
-
-class TradeInsightAiRendering(TradeInsightAiBase):
-    disclaimer: str
-    card_order: list[str] = Field(default_factory=list)
-
-
-class TradeInsightAiGuardrails(TradeInsightAiBase):
-    statuses_preserved: bool
-    risk_flags_preserved: bool
-    no_executable_recommendations: bool
-
-
-class TradeInsightAiOutcome(TradeInsightAiBase):
-    schema_version: str
-    analysis_produced_at: datetime
-    ticker: str
-    underlying_price: str | None = None
-    snapshot: TradeInsightAiSnapshotMeta
-    headline: TradeInsightAiHeadline
-    metric_cards: list[TradeInsightAiMetricCard]
-    scenario_cards: list[TradeInsightAiScenarioCard]
-    score_breakdown: list[TradeInsightAiScoreBreakdown]
-    section_cards: TradeInsightAiSectionCards
-    vrp_assessment: TradeInsightAiVrpAssessment | None = None
-    preferred_expression: TradeInsightAiPreferredExpression | None = None
-    dominant_read: TradeInsightAiDominantRead
-    best_expressions: list[TradeInsightAiBestExpression] = Field(default_factory=list)
-    conflicts: list[TradeInsightAiConflict] = Field(default_factory=list)
-    required_checks: list[TradeInsightAiRequiredCheck] = Field(default_factory=list)
-    rejected_ideas: list[TradeInsightAiRejectedIdea] = Field(default_factory=list)
-    missing_data: list[str] = Field(default_factory=list)
-    rendering: TradeInsightAiRendering
-    guardrails: TradeInsightAiGuardrails
-
-
-class TradeInsightAiAnalysisRequest(TradeInsightAiBase):
-    force_rerun: bool = False
-
-
-class TradeInsightAiAnalysisResponse(TradeInsightAiBase):
-    analysis_id: UUID
-    ticker: str
-    run_id: int
-    trade_insights_input_hash: str
-    analysis_input_hash: str
-    model: str
-    prompt_version: str
-    status: Literal["queued", "running", "succeeded", "failed"]
-    produced_at: datetime | None = None
-    outcome: TradeInsightAiOutcome | None = None
-    markdown: str | None = None
-    error_message: str | None = None
-    requested_at: datetime
-    started_at: datetime | None = None
-    finished_at: datetime | None = None
-    reused: bool = False
 
 
 # ---------------------------------------------------------------------------
